finish migrating all former main configuration stuff to better homes
The studio integration stuff now finds itself in the apollo_telemetry
module which is a good home for it.

The opentelemetry stuff now finds itself in the opentelemetry plugin,
which, again, is a good home for it.
